Support root domain records
We weren't able to perform any work on root domain records as the API
wants us to omit the "name" parameter while the module requires one. By
supporting the '@' name we can service such records. In addition, the
update_record() method needed to switch to using a record ID to support
this.

The added logic warrants a small internal function to format the payload
as well.

Fixes #4